requirements.txt
Flask
python-dotenv
requests
PyMuPDF
faiss-cpu
sentence-transformers
langchain
